requirements.txt
django>=3.0.7,<3.1.0
djangorestframework>=3.11.0,<3.12.0

django-phone-field>=1.8.1,<1.9.0

psycopg2>=2.8.5,<2.9.0

flake8>=3.8.3,<3.9.0

djangorestframework_simplejwt>=4.4.0,<4.5.0